Refactor and fix purpose detection issue
When purpose detection is enabled, optionsSettingsHandler first ensures the HTTP request didn't fail and expects a response; otherwise, all purpose checks are skipped. Because of this condition, all important subdomain purpose checks were generally ignored. This issue has now been fixed by removing the status code condition.
