+++ internal/coreutils/analysis/run.go
@@ -1,0 +1,12 @@
+package analysis
+
+func (check *SubdomainCheck) Misconfigurations() {
+	check.hostHeaders()         // Host header injections
+	check.cookieInjectionPath() // session hijacking, xss
+}
+
+func (check *SubdomainCheck) Purpose() {
+	check.mailServer() // General, Exchange
+	check.api()        // Content types, API versions, rate limit
+	check.login()      // Scan response body for login indicators
+}
