+++ internal/streams/output.go
@@ -116,9 +116,8 @@ func optionsSettingsHandler(settings shared.SettingsHandler) bool {
 		utils.PingWrapper(settings.ConsoleOutput, settings.Params.Subdomain, settings.Args.PingCount)
 	}
 	requests.SetDnsEnumType() // Handle type by global switch
-	// httpCodeCheck: do not perform analysis if the HTTP request fails (-1)
-	if settings.Args.DetectPurpose && requests.HttpCodeCheck(settings, url) {
-		settings.ConsoleOutput.WriteString(" | Trying To Identify The Subdomain Purpose...\n")
+	if settings.Args.DetectPurpose {
+		settings.ConsoleOutput.WriteString(" | Trying to identify the subdomain purpose...\n")
 		shared.GShowAllHeaders = true
 		headers := requests.AnalyseHttpHeader(settings.HttpClient, settings.Params.Subdomain, settings.Args.HttpRequestMethod)
 		check := analysis.SubdomainCheck{
@@ -127,20 +126,17 @@ func optionsSettingsHandler(settings shared.SettingsHandler) bool {
 			HttpHeaders:   headers,
 			HttpClient:    settings.HttpClient,
 		}
-		check.MailServer()
-		check.API()
-		check.Login()
+		check.Purpose() // run.go
 	}
+	// httpCodeCheck: do not perform analysis if the HTTP request fails (-1)
 	if settings.Args.MisconfTest && requests.HttpCodeCheck(settings, url) {
-		settings.ConsoleOutput.WriteString(" | Testing for weaknesses...\n")
+		settings.ConsoleOutput.WriteString(" | Testing for common weaknesses...\n")
 		check := analysis.SubdomainCheck{
 			Subdomain:     settings.Params.Subdomain,
 			ConsoleOutput: settings.ConsoleOutput,
 			HttpClient:    settings.HttpClient,
 		}
-		check.CORS()
-		check.HeaderInjection()
-		// ...
+		check.Misconfigurations() // run.go
 	}
 	return true
 }
